Crate cirrus_test_service

Crate used for testing with Cirrus.

Re-exports

pub use cirrus_test_runtime as runtime;

Modules

chain_spec

Chain specification for the cirrus test runtime.

Structs

RuntimeExecutor

Native executor instance.

TestNode

A Cumulus test node instance used for testing.

TestNodeBuilder

A builder to create a TestNode.

Enums

Keyring

Set of test accounts.

Functions

construct_extrinsic

Construct an extrinsic that can be applied to the test runtime.

new_partial

Starts a ServiceBuilder for a full service.

node_config

Create a Cumulus Configuration.

run_primary_chain_validator_node

Run a primary-chain validator node.

Type Definitions

Backend

The backend type used by the test service.

Client

The client type being used by the test service.

CodeExecutor

Code executor for the test service.

Executor

Secondary executor for the test service.

WrapAnnounceBlockFn

The signature of the announce block fn.